infra+caaas+infer: clearml bind-mount mode + fp16 inference + show_tab
- submit_clearml_task.sh: keep bind-mount mode (-v /home/hfunaya/git/
  e2e_calib:/workspace) + cd /workspace + one-liner bash_setup_script.
  Tried --folder upload mode but agent v3 docker mode does not chdir
  into the unpacked folder and joins multi-line setup with `;`, both
  break. Stay with the legacy rsync+bind-mount path.
- infer_pipeline.infer_one: float16 autocast (bf16 silently fp32
  emulates on V100 sm_70 and shifts Δuv distribution).
- test_ba_smoke: target current ckpt + dump GT pert_vec for diff.
- caaas_app: add show_tab() js function so the Tile/Full tab buttons
  actually switch.

+++ scripts/inference/infer_pipeline.py
@@ -96,7 +96,10 @@ def infer_one(model, ds, idx: int, device: str = 'cuda', seed: int | None = None
         point_in = torch.cat([dist_uvd[..., :3], dist_uvd[..., 4:5]], dim=-1)
     else:
         point_in = dist_uvd[..., :3]
-    with torch.autocast(device_type='cuda', dtype=torch.bfloat16):
+    # fp16 to match training (--mixed_precision=fp16). bf16 silently runs as
+    # fp32-emulated on V100 (sm_70) and shifts the per-pt Δuv distribution
+    # away from what the model actually saw during training.
+    with torch.autocast(device_type='cuda', dtype=torch.float16):
         out = model(img_in,
                     point_in.unsqueeze(0).to(device),
                     key_padding_mask=pad,
